requirements.txt
connexion
flask
flask-restful
mysql-connector
